requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.0.0
gpxpy>=1.5.0
geopandas>=0.13.0
contextily>=1.3.0
matplotlib>=3.7.0
PySide6>=6.5.0
tqdm>=4.65.0
numpy>=1.24.0